Backend/requirements.txt
# Web Framework
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
pymongo==4.6.1
mongoengine==0.27.0
dnspython==2.4.2

# Authentication
firebase-admin==6.3.0
python-dotenv==1.0.0

# Browser Automation
selenium==4.16.0
webdriver-manager==4.0.1

# Audio Recording
soundcard==0.4.2
numpy==1.26.4
sounddevice==0.4.6
scipy==1.12.0
wavio==0.0.8

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Utils
python-dateutil==2.8.2
pywin32==306; platform_system == "Windows"

# Google Cloud
google-cloud-speech==2.25.1
google-cloud-storage==2.14.0
